requirements.txt
kafka-python
requests
psycopg2-binary
aiokafka 
